Refactoring storj.js
